@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.1'
     defaultConfig {
         applicationId "com.myapp.lexicon"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionName 'beta'
         versionCode 32
         multiDexEnabled true
@@ -23,37 +23,41 @@ android {
         checkReleaseBuilds false
         abortOnError false
     }
+    compileOptions {
+        sourceCompatibility = 1.8
+        targetCompatibility = 1.8
+    }
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'com.google.android.material:material:1.1.0'
     implementation 'com.sothree.slidinguppanel:library:3.3.1'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.google.firebase:firebase-core:16.0.8'
-    implementation 'com.google.firebase:firebase-storage:16.1.0'
-    implementation 'com.google.firebase:firebase-ads:17.2.0'
-    implementation 'com.google.firebase:firebase-appindexing:17.1.0'
-    implementation 'com.google.android.gms:play-services-ads:17.2.0'
-    implementation 'com.google.android.gms:play-services-plus:16.0.0'
-    implementation 'com.google.android.gms:play-services-auth:16.0.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.firebase:firebase-core:17.2.2'
+    implementation 'com.google.firebase:firebase-storage:19.1.1'
+    implementation 'com.google.firebase:firebase-ads:18.3.0'
+    implementation 'com.google.firebase:firebase-appindexing:19.1.0'
+    implementation 'com.google.android.gms:play-services-ads:18.3.0'
+    implementation 'com.google.android.gms:play-services-plus:17.0.0'
+    implementation 'com.google.android.gms:play-services-auth:17.0.0'
     testImplementation 'junit:junit:4.12'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'com.android.volley:volley:1.1.1'
-    implementation 'com.android.support:customtabs:28.0.0'
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
+    implementation 'androidx.browser:browser:1.2.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 
 //    implementation "android.arch.persistence.room:runtime:1.1.1"
 //    annotationProcessor "android.arch.persistence.room:compiler:1.1.1"
 
-    implementation 'io.reactivex.rxjava2:rxjava:2.0.8'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.10'
     implementation 'com.jakewharton.rxbinding2:rxbinding:2.0.0'
     implementation 'org.greenrobot:eventbus:3.1.1'
-    implementation 'com.android.support:percent:28.0.0'
+    implementation 'androidx.percentlayout:percentlayout:1.0.0'
 
-    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'androidx.multidex:multidex:2.0.1'
 
 }
 
