@@ -4,7 +4,7 @@ import com.myapp.lexicon.database.DataBaseEntry;
 
 public class TranslateDialogEvent
 {
-    public final DataBaseEntry entry;
+    final DataBaseEntry entry;
 
     public TranslateDialogEvent(DataBaseEntry entry)
     {
